feat: split up typedef
